+++ sdk/network/azure-mgmt-frontdoor/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 2.0.0 (2025-04-25)
+
+### Features Added
+
+  - Enum `ActionType` added member `CAPTCHA`
+  - Enum `MatchProcessingBehavior` added member `CONTINUE`
+  - Model `PolicySettings` added property `captcha_expiration_in_minutes`
+
+### Breaking Changes
+
+  - Deleted or renamed enum value `MatchProcessingBehavior.CONTINUE_ENUM`
+
 ## 1.2.0 (2024-04-15)
 
 ### Features Added